✨ Initial 1.12 commit
Due to an irreversible mistake made upon returning to mod development to
update JEI Integration, the repository had to be recreated. As a result,
all commit history prior to 1.15 is no longer present.

This update (1.6.0) brings a number of changes, fixes and improvements
all of which are listed below:

- Update Forge to 14.23.5.2847
- Update MCP mappings to stable_39
- Update JEI to 4.15.0+
- Update CurseGradle to 1.4.0
- Fix JEI dependency slug in curseforge build script task
- Replace requiredLibrary with non-deprecated method requiredDependency
in curseforge build script task
- Update repo URL in gradle.properties
- Update LICENSE copyright year (inc. source code file headers)
- Update mcmod.info
- Update logo (mcmod.info)
- Fix broken method calls in TooltipHandler due to MCP mapping changes
- Update README.md
- Fix Travis CI build script
- Remove .github/CONTRIBUTING.md (unused)
- Bump version to 1.6.0